requirements.txt
numpy==1.23.4
pandas==1.5.1
Pillow==9.4.0
Plost==0.2.5
plotly==5.13.0
streamlit==1.17.0
